core: Add SPCI support for a UP OP-TEE runtime
This patch adds support for using the SPCI messaging framework during OP-TEE
runtime. Instead of using private SMCs, SPCI RX/TX buffers are used for message
passing. The SPCI hooks replace the private SMC invocations to ensure that the
core OP-TEE functionality is preserved and only the "transport" is changed.

This patch only enables runtime support on a single CPU i.e. it is not MP or
thread-safe. Also, it is assumed that the power management hooks are not
used. These features will be added in later patches.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>